Cleanup in trackcanvas
Removed out commented printfs, unused parts of source
Implemented posnum. A lot of cosmetics

+++ cinelerra/edithandles.C
@@ -133,7 +133,7 @@ void EditHandles::update()
 		for(Edit *edit = current->edits->first; edit; edit = edit->next)
 		{
 // Test in point
-			int64_t handle_x, handle_y, handle_w, handle_h;
+			int handle_x, handle_y, handle_w, handle_h;
 			trackcanvas->get_handle_coords(edit, handle_x, handle_y, handle_w, handle_h, EDIT_IN);
 
 			if(visible(handle_x, handle_y, handle_w, handle_h))
